requirements.txt
python-telegram-bot
google-generativeai
google-cloud-texttospeech
python-dotenv